# openapi_apidog.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Bimbi AI API",
    "description": "API collection documentation for Bimbi AI Backend. Built with Clean Architecture using Go and Gin. Includes authentication, RAG-based insights generation, longitudinal child tracking, and home activity recommendations with done-status tracking.",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Development Server"
    },
    {
      "url": "http://bimbi-be-dev.btdev.my.id",
      "description": "Development Environment"
    }
  ],
  "paths": {
...
    "/api/assessments": {
      "get": {
        "summary": "Get Assessments by Child",
        "description": "Get all assessments for a specific child (raw DB records with input_payload and ai_response as JSONB).",
        "operationId": "getAssessments",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "child_id",
            "in": "query",
            "required": true,
            "description": "Child UUID",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Assessment" }
                },
                "example": [
                  {
                    "id": "712deb8a-da13-46e3-b36e-da27d7f05a17",
                    "child_id": "123e4567-e89b-12d3-a456-426614174000",
                    "assessment_date": "2026-06-11T09:57:32Z",
                    "input_payload": "{\"child_id\":\"123e4567-e89b-12d3-a456-426614174000\",\"daily_activities\":[\"menggambar\",\"bermain balok\"],\"parent_anxiety\":\"Anak susah fokus\",\"positive_triggers\":\"Antusias saat membangun sesuatu\",\"parent_goals\":\"Ingin tahu cara mendampingi belajar\"}",
                    "ai_response": "{\"talent_label\":\"Kecerdasan Spasial-Visual\",\"empathetic_analysis\":\"...\",\"theoretical_basis\":\"...\",\"home_activities\":[\"...\"],\"learning_hacks\":[\"...\"],\"sources\":[\"multiple_intelligences_gardner.pdf\"]}",
                    "created_at": "2026-06-11T09:57:58Z"
                  }
                ]
              }
            }
          },
          "400": {
            "description": "Missing child_id query parameter",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "bad_request", "message": "child_id query parameter is required" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "unauthorized", "message": "Missing or invalid token" }
              }
            }
          },
          "403": {
            "description": "Forbidden - child does not belong to this user",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "forbidden", "message": "Access denied" }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Generate Insights (New Assessment)",
        "description": "Submit a parent's observations about their child to receive an AI-generated empathetic talent analysis, home activity recommendations, and learning hacks. The result is saved as a new Assessment record. Previous assessment + feedback history is automatically used to generate `progress_analysis` and improve `home_activities`.",
        "operationId": "generateInsights",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AssessmentRequest" },
              "example": {
                "child_id": "123e4567-e89b-12d3-a456-426614174000",
                "daily_activities": ["menggambar", "bermain balok", "menonton video edukasi"],
                "parent_anxiety": "Anak saya susah fokus dan selalu bergerak, saya khawatir dia tidak bisa belajar dengan baik di sekolah.",
                "positive_triggers": "Sangat antusias saat diajak membangun sesuatu dari balok atau kardus bekas.",
                "parent_goals": "Ingin tahu cara terbaik mendampingi anak belajar di rumah sesuai karakternya."
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Insights generated successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InsightResponse" },
                "example": {
                  "talent_label": "Kecerdasan Spasial-Visual",
                  "empathetic_analysis": "Kekhawatiran Anda sangat wajar dan banyak orang tua merasakannya. Anak yang terus bergerak seringkali bukan tanda masalah, melainkan tanda energi kognitif yang tinggi.\n\nBerdasarkan aktivitas seperti membangun balok dan menggambar, Ara menunjukkan kecerdasan Spasial-Visual yang kuat. Kemampuan ini adalah fondasi bagi profesi seperti arsitek, desainer, dan insinyur.",
                  "theoretical_basis": "Berdasarkan teori Kecerdasan Majemuk (Multiple Intelligences) Howard Gardner, kecerdasan spasial-visual ditandai dengan kemampuan anak untuk berpikir dalam tiga dimensi dan merepresentasikan dunia secara visual.",
                  "progress_analysis": "Dibandingkan bulan lalu, Ara kini lebih mampu fokus pada aktivitas konstruksi dan menyelesaikannya hingga tuntas. Ini adalah perkembangan yang luar biasa, Anda sudah melakukan hal yang tepat!",
                  "home_activities": [
                    "Bangun 'kota mini' dari kardus bekas sabun dan botol shampoo bersama Ara selama 15 menit sebelum tidur.",
                    "Ajak Ara mendekorasi tempat belajarnya sendiri dengan gambar-gambar yang dia buat.",
                    "Sediakan plastisin atau tanah liat untuk membuat bentuk hewan favoritnya."
                  ],
                  "learning_hacks": [
                    "Saat menjelaskan sesuatu, gunakan gambar atau sketsa sederhana, bukan hanya kata-kata.",
                    "Beri Ara 'tugas visual': minta dia menggambar apa yang baru dipelajari sebagai pengganti menulis ringkasan."
                  ],
                  "sources": ["multiple_intelligences_gardner.pdf", "learning_styles_vak.pdf"]
                }
              }
            }
          },
          "400": {
            "description": "Invalid payload",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "error": "invalid_payload",
                  "message": "Invalid request body: Key: 'AssessmentRequest.ChildID' Error:Field validation for 'ChildID' failed on the 'required' tag"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "unauthorized", "message": "Missing or invalid token" }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "service_error", "message": "llm error: context deadline exceeded" }
              }
            }
          }
        }
      }
    },
    "/api/assessments/{id}": {
      "get": {
        "summary": "Get Assessment by ID",
        "description": "Fetch a single assessment record (raw DB object) by its UUID.",
        "operationId": "getAssessment",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Assessment UUID",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Assessment" },
                "example": {
                  "id": "712deb8a-da13-46e3-b36e-da27d7f05a17",
                  "child_id": "123e4567-e89b-12d3-a456-426614174000",
                  "assessment_date": "2026-06-11T09:57:32Z",
                  "input_payload": "{\"child_id\":\"123e4567-e89b-12d3-a456-426614174000\",\"daily_activities\":[\"menggambar\",\"bermain balok\"],\"parent_anxiety\":\"Anak susah fokus\",\"positive_triggers\":\"Antusias saat membangun sesuatu\",\"parent_goals\":\"Ingin tahu cara mendampingi belajar\"}",
                  "ai_response": "{\"talent_label\":\"Kecerdasan Spasial-Visual\",\"empathetic_analysis\":\"...\",\"home_activities\":[\"...\"],\"sources\":[\"multiple_intelligences_gardner.pdf\"]}",
                  "created_at": "2026-06-11T09:57:58Z"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - assessment does not belong to this user's child",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "forbidden", "message": "Access denied" }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "not_found", "message": "Assessment not found" }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Assessment",
        "operationId": "deleteAssessment",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Assessment UUID",
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "message": { "type": "string" } }
                },
                "example": { "message": "Assessment deleted successfully" }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "forbidden", "message": "Access denied" }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "not_found", "message": "Assessment not found" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter the JWT token obtained from the `/api/auth/login` endpoint."
      }
    },
    "schemas": {
...
      "Assessment": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid", "example": "712deb8a-da13-46e3-b36e-da27d7f05a17" },
          "child_id": { "type": "string", "format": "uuid", "example": "123e4567-e89b-12d3-a456-426614174000" },
          "assessment_date": { "type": "string", "format": "date-time", "example": "2026-06-11T09:57:32Z" },
          "input_payload": {
            "type": "string",
            "description": "Raw JSON string of the original request payload",
            "example": "{\"child_id\":\"123e4567-e89b-12d3-a456-426614174000\",\"daily_activities\":[\"menggambar\",\"bermain balok\"],\"parent_anxiety\":\"Anak susah fokus\",\"positive_triggers\":\"Antusias saat membangun sesuatu\",\"parent_goals\":\"Ingin tahu cara mendampingi belajar\"}"
          },
          "ai_response": {
            "type": "string",
            "description": "Raw JSON string of the AI-generated insights",
            "example": "{\"talent_label\":\"Kecerdasan Spasial-Visual\",\"empathetic_analysis\":\"...\",\"theoretical_basis\":\"...\",\"home_activities\":[\"...\"],\"learning_hacks\":[\"...\"],\"sources\":[\"multiple_intelligences_gardner.pdf\"]}"
          },
          "created_at": { "type": "string", "format": "date-time", "example": "2026-06-11T09:57:58Z" }
        }
      },
      "AssessmentRequest": {
        "type": "object",
        "required": ["child_id", "daily_activities", "parent_anxiety", "positive_triggers", "parent_goals"],
        "properties": {
          "child_id": {
            "type": "string",
            "format": "uuid",
            "description": "UUID of the child profile",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "daily_activities": {
            "type": "array",
            "items": { "type": "string" },
            "description": "List of activities the child frequently does at home",
            "example": ["menggambar", "bermain balok", "menonton video edukasi"]
          },
          "parent_anxiety": {
            "type": "string",
            "description": "The parent's main concern or worry about their child",
            "example": "Anak saya susah fokus dan selalu bergerak, saya khawatir dia tidak bisa belajar dengan baik di sekolah."
          },
          "positive_triggers": {
            "type": "string",
            "description": "Situations or topics that make the child visibly excited and engaged",
            "example": "Sangat antusias saat diajak membangun sesuatu dari balok atau kardus bekas."
          },
          "parent_goals": {
            "type": "string",
            "description": "What the parent hopes to achieve or understand about their child",
            "example": "Ingin tahu cara terbaik mendampingi anak belajar di rumah sesuai karakternya."
          }
        }
      },
      "InsightResponse": {
        "type": "object",
        "properties": {
          "talent_label": {
            "type": "string",
            "description": "Kecerdasan utama anak berdasarkan teori Multiple Intelligences",
            "example": "Kecerdasan Spasial-Visual"
          },
          "empathetic_analysis": {
            "type": "string",
            "description": "Dua paragraf: validasi kekhawatiran + bingkai ulang sebagai potensi bakat",
            "example": "Kekhawatiran Anda sangat wajar...\n\nBerdasarkan aktivitas seperti membangun balok..."
          },
          "theoretical_basis": {
            "type": "string",
            "description": "Landasan teori psikologi atau pendidikan yang mendasari analisis",
            "example": "Berdasarkan teori Kecerdasan Majemuk Howard Gardner..."
          },
          "progress_analysis": {
            "type": "string",
            "description": "Perbandingan perkembangan dari asesmen sebelumnya (hanya ada jika ada asesmen sebelumnya)",
            "example": "Dibandingkan bulan lalu, Ara kini lebih mampu fokus pada aktivitas konstruksi."
          },
          "home_activities": {
            "type": "array",
            "items": { "type": "string" },
            "description": "3 rekomendasi aktivitas rumah yang spesifik",
            "example": [
              "Bangun 'kota mini' dari kardus bekas sabun dan botol shampoo bersama Ara selama 15 menit sebelum tidur.",
              "Ajak Ara mendekorasi tempat belajarnya sendiri dengan gambar-gambar yang dia buat.",
              "Sediakan plastisin atau tanah liat untuk membuat bentuk hewan favoritnya."
            ]
          },
          "learning_hacks": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Tips praktis mendampingi gaya belajar anak di rumah",
            "example": [
              "Saat menjelaskan sesuatu, gunakan gambar atau sketsa sederhana, bukan hanya kata-kata.",
              "Beri Ara 'tugas visual': minta dia menggambar apa yang baru dipelajari sebagai pengganti menulis ringkasan."
            ]
          },
          "sources": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Nama file PDF sumber konteks RAG yang digunakan",
            "example": ["multiple_intelligences_gardner.pdf", "learning_styles_vak.pdf"]
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "invalid_payload" },
          "message": { "type": "string", "example": "Key: 'Field' Error:Field validation failed" }
        }
      },
...
    }
  }
}